// app/models/Question.ts
export type Question = {
  id: number
  numero: number
  texto: string
  examen: number
  correcta: number
  tema: number
  ano: number
  opciones: number
  opcion1: string
  opcion2: string
  opcion3: string
  opcion4: string
  opcion5: string
  clasifica1: number
  clasifica2: number
  clasifica3: number
  clasifica4: number
  clasifica5: number
  comentario: string
  image: string
}